+++ lint-compose/src/main/kotlin/team/duckie/quackquack/lint/compose/empty.kt
@@ -1,0 +1,10 @@
+/*
+ * Designed and developed by 2022 SungbinLand, Team Duckie
+ *
+ * [empty.kt] created by Ji Sungbin on 22. 8. 14. 오전 12:57
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
+ */
+
+package team.duckie.quackquack.lint.compose
