+++ lint-quack/build.gradle.kts
@@ -1,0 +1,12 @@
+/*
+ * Designed and developed by 2022 SungbinLand, Team Duckie
+ *
+ * [build.gradle.kts] created by Ji Sungbin on 22. 8. 18. 오후 4:30
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
+ */
+
+plugins {
+    id(PluginEnum.Lint)
+}
