+++ lint-quack/src/main/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Designed and developed by 2022 SungbinLand, Team Duckie
+  ~
+  ~ [AndroidManifest.xml] created by Ji Sungbin on 22. 8. 18. 오후 4:30
+  ~
+  ~ Licensed under the MIT.
+  ~ Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
+  -->
+
+<manifest />
