+++ lint-quack/src/main/resources/META-INF/services/com.android.tools.lint.client.api.IssueRegistry
@@ -7,4 +7,4 @@
 # Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
 #
 
-team.duckie.quackquack.lint.compose.DesignSystemIssueRegistry
+team.duckie.quack.lint.quack.DesignSystemIssueRegistry
